Notification system, searching a task
Notification system is implemented. For now, it notifies tasks only when their time comes.(There is no other type notifications) Notification system works properly for only list owners for now.(It is not working for shared person)
Search task functionality is started to implement in the front side of the project.